Экспорт в эксель и проверка на права администратора сделано

+++ ServerME/Services/LogService.cs
@@ -8,17 +8,26 @@ namespace ServerME.Services
     public class LogService
     {
         private LogRepository repository;
+        private PrivilegeService privilegeService;
         public LogService()
         {
             repository = new LogRepository();
+            privilegeService = new PrivilegeService();
         }
         public List<string[]> GetLogs(string filter, string sorting, int currentPage, int pageSize, User user)
         {
-            var privilege = new Dictionary<string, string>();
-            
-            var logs = repository.GetAnimals(filter, sorting, privilege, currentPage, pageSize);
-            var mappedLogs = MapLogs(logs);
-            return mappedLogs;
+            var resultCheck = privilegeService.CheckUserForLogs(user);
+
+            if (resultCheck)
+            {
+                var logs = repository.GetLogs(filter, sorting, currentPage, pageSize);
+                var mappedLogs = MapLogs(logs);
+                return mappedLogs;
+            }
+            else
+            {
+                throw new InvalidOperationException("У вас нет прав администрирования!");
+            }
         }
 
         private List<string[]> MapLogs(IEnumerable<Log> logs)
@@ -50,9 +59,37 @@ namespace ServerME.Services
             return result;
         }
 
+        public byte[] GetExcelByteArrayFormat(string filter, string sorting, User user)
+        {
+            var resultCheck = privilegeService.CheckUserForLogs(user);
+
+            if (resultCheck)
+            {
+                string[] columnNames = new string[] {"Операция", "ФИО", "Телефон", "Эл. почта", "Организация",
+                "Структурное подразделение", "Должность", "Раб. телефон", "Раб. адрес эл. почта",
+                "Логин", "Дата и время", "Объект", "ID объекта",  "Описание объекта после совершения действия", "Загруженный файл"};
+
+                var logs = GetLogs(filter, sorting, 1, int.MaxValue, user);
+                return ExcelConverter.GenerateExcelFile(logs, columnNames);
+            }
+            else
+            {
+                throw new InvalidOperationException("У вас нет прав администрирования!");
+            }
+        }
+
         public void DeleteLogs(IEnumerable<int> logsIdArray, User user)
         {
-            repository.DeleteLogs(logsIdArray);
+            var resultCheck = privilegeService.CheckUserForLogs(user);
+
+            if (resultCheck)
+            {
+                repository.DeleteLogs(logsIdArray);
+            }
+            else
+            {
+                throw new InvalidOperationException("У вас нет прав администрирования!");
+            }
         }
     }
 }
